Rewrote implementation using go-kit.
Rewrote all functionality in order to separate in well defined layers.
Now we have a Model, a Data Access layer, a Service layer, the Endpoints layer, and the Transport layer.
From this experience, it seems like we could separate all layers until the Endpoint layer in a module, and implement the Transport layer in the application.